modularized ingestion script
